refactor: sale 패키지 이동
