+++ domain/domain-pos/src/main/java/domain/pos/order/service/OrderService.java
@@ -28,7 +28,7 @@ import domain.pos.receipt.entity.Receipt;
 import domain.pos.receipt.implement.ReceiptCustomerWriter;
 import domain.pos.receipt.implement.ReceiptReader;
 import domain.pos.receipt.implement.ReceiptValidator;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.store.implement.SaleReader;
 import domain.pos.store.implement.SaleValidator;
 import lombok.RequiredArgsConstructor;
