+++ domain/domain-pos/src/main/java/domain/pos/receipt/entity/Receipt.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import domain.pos.order.entity.Order;
 import domain.pos.order.entity.vo.OrderStatus;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.table.entity.Table;
 import lombok.Getter;
