+++ domain/domain-pos/src/main/java/domain/pos/receipt/implement/ReceiptWriter.java
@@ -10,7 +10,7 @@ import com.exception.ServiceException;
 
 import domain.pos.receipt.entity.Receipt;
 import domain.pos.receipt.repository.ReceiptRepository;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.table.entity.Table;
 import lombok.RequiredArgsConstructor;
