+++ domain/domain-pos/src/main/java/domain/pos/receipt/repository/ReceiptRepository.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 
 import domain.pos.receipt.entity.Receipt;
 import domain.pos.receipt.entity.ReceiptInfo;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.table.entity.Table;
 
 @Repository
