+++ domain/domain-pos/src/main/java/domain/pos/store/service/SaleService.java
@@ -9,7 +9,7 @@ import com.exception.ServiceException;
 import com.vo.UserPassport;
 
 import domain.pos.receipt.implement.ReceiptReader;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.store.entity.Store;
 import domain.pos.store.implement.SaleReader;
 import domain.pos.store.implement.SaleWriter;
