+++ infra/rdb/pos/src/main/java/com/pos/receipt/mapper/ReceiptMapper.java
@@ -12,7 +12,7 @@ import com.pos.table.entity.TableEntity;
 import domain.pos.order.entity.Order;
 import domain.pos.receipt.entity.Receipt;
 import domain.pos.receipt.entity.ReceiptInfo;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.table.entity.Table;
 
 public class ReceiptMapper {
