+++ infra/rdb/pos/src/test/java/com/pos/call/repository/CallRepositoryImplTest.java
@@ -35,7 +35,7 @@ import domain.pos.call.entity.Call;
 import domain.pos.call.entity.CallMessage;
 import domain.pos.call.repository.CallRepository;
 import domain.pos.receipt.entity.Receipt;
-import domain.pos.store.entity.Sale;
+import domain.pos.sale.entity.Sale;
 import domain.pos.store.entity.Store;
 
 class CallRepositoryImplTest extends RepositoryTest {
